requirements.txt
# Updated dependencies for Python 3.13 compatibility (January 2025)
atproto==0.0.61
python-dotenv==1.1.0
# Use latest psycopg2-binary with Python 3.13 support
psycopg2-binary==2.9.10
# Use psycopg (modern async PostgreSQL driver) instead of asyncpg for Python 3.13 compatibility
psycopg[binary]==3.2.4
httpx==0.28.1
pydantic==2.11.4
cbor2==5.6.0
websockets==13.1
psutil==6.1.1
colorama==0.4.6
# asyncio is part of the standard library - no need to install separately
# Using built-in atproto firehose